// packages/protocols/pools/src/model.generated.ts
// Generated by scripts/generate/generate-pools-package.ts. Do not edit.
// sha256:4dcc41e9c3b182ed8afd6583066fb4300bfb013eec6dff9cfe228575663efa99
export const POOL_MODEL = {
  cl: {
    constants: {
      Q96: "79228162514264337593543950336",
      Q128: "340282366920938463463374607431768211456",
      minTick: -887272,
      maxTick: 887272,
      minSqrtRatioX96: "4295128739",
      maxSqrtRatioX96: "1461446703485210103287273052203988822378723970342",
    },
    tickMultipliers: [
      "0xfffcb933bd6fad37aa2d162d1a594001",
      "0xfff97272373d413259a46990580e213a",
      "0xfff2e50f5f656932ef12357cf3c7fdcc",
      "0xffe5caca7e10e4e61c3624eaa0941cd0",
      "0xffcb9843d60f6159c9db58835c926644",
      "0xff973b41fa98c081472e6896dfb254c0",
      "0xff2ea16466c96a3843ec78b326b52861",
      "0xfe5dee046a99a2a811c461f1969c3053",
      "0xfcbe86c7900a88aedcffc83b479aa3a4",
      "0xf987a7253ac413176f2b074cf7815e54",
      "0xf3392b0822b70005940c7a398e4b70f3",
      "0xe7159475a2c29b7443b29c7fa6e889d9",
      "0xd097f3bdfd2022b8845ad8f792aa5825",
      "0xa9f746462d870fdf8a65dc1f90e061e5",
      "0x70d869a156d2a1b890bb3df62baf32f7",
      "0x31be135f97d08fd981231505542fcfa6",
      "0x9aa508b5b7a84e1c677de54f3e99bc9",
      "0x5d6af8dedb81196699c329225ee604",
      "0x2216e584f5fa1ea926041bedfe98",
      "0x48a170391f7dc42444e8fa2",
    ],
    sourceDigest: "caf37d210361f0141f7e19b78581e634bbc186763ecadd8be9f378d95892386a",
    swapFeeScale: "1000000",
    swapSourceDigest: "234e487f971ccd2588d282cccffd6306f667643c842891378fcf9a867954f072",
  },
  feeIndexScale: "1000000000000000000",
  swapFeeDenominator: "10000",
  musdc: {
    address: "0x04671c72aab5ac02a03c1098314b1bb6b560c197",
    implementationAddress: "0x09d42a1c4792546c26451af2d373bf0d4c545baf",
    implementationSlot: "0x360894a13ba1a3210667c828492db98dca3e2076cc3735a920a3ca505d382bbc",
    addressCodeSha256: "4bc884c60f64f3486dd8c1a31ede2542f7417edc9e94730d9a86ae386cb82f65",
    implementationCodeSha256: "378d40bcb07169de058746028e58f77128dcde6ff1660a70dbe338a7e700abc2",
  },
  musdt: {
    address: "0xeb5a5d39de4ea42c2aa6a57eca2894376683bb8e",
    decimals: 6,
    implementationAddress: "0xa9187635a3f6b245888c874e57dbd9fc7139864d",
    implementationSlot: "0x360894a13ba1a3210667c828492db98dca3e2076cc3735a920a3ca505d382bbc",
    addressCodeSha256: "74d711981a42c17959bd1cff5b365bf85f36e5fe11d2a4649ea437df976bce96",
    implementationCodeSha256: "fdc02144e01e0db4d7179d46cf6e51edbda2d9aacf0ee83f1700788be80f8cef",
  },
} as const;